GetSingleEmployee fixed

+++ EmployeeTrackingWebApi/Contracts/ShiftResponse.cs
@@ -5,6 +5,5 @@ public class ShiftResponse
     public int Id { get; set; }
     public DateTime? StartShift { get; set; }
     public DateTime? EndShift { get; set; }
-    public int TotalHoursWorked { get; set; }
-  
+    public double TotalHoursWorked { get; set; }
 }
