+++ EmployeeTrackingWebApi/Models/Shifts.cs
@@ -5,7 +5,7 @@ public class Shifts
     public int Id { get; set; }
     public DateTime? StartShift { get; set; }
     public DateTime? EndShift { get; set; }
-    public int TotalHoursWorked { get; set; }
+    public double TotalHoursWorked { get; set; }
     public Employee Employee { get; set; }
     
 }
