+++ EmployeeTrackingWebApi/Services/IShiftService.cs
@@ -4,5 +4,5 @@ public interface IShiftService
 {
     DateTime StartShift(int id);
     DateTime EndShift(int id);
-    int totalHoursWorked();
+    double totalHoursWorked();
 }
